Cleaning up Test Projects, consolidating it to get unified coverage data.
