--- SmugMugCore.Net/Data20/MethodDetailsInfo.cs
@@ -1,10 +1,0 @@
-using System.Text.Json.Serialization; 
-
-namespace SmugMugCore.Net.Data20
-{
-    public class MethodDetailsInfo
-    {
-        [JsonPropertyName("Permissions")]
-        public List<string>? Permissions { get; set; } // [1]
-    }
-}
