+++ TestSmugMugCore/Content/TestBmpImage.bmp
