+++ TestSmugMugCore/Content/TestJpegImage.jpeg
