+++ TestSmugMugCore/Content/TestPngImage.png
