+++ TestSmugMugCore/Content/TestScannedImage.JPG
