+++ TestSmugMugCore/TestSmugMugCore.csproj
@@ -62,6 +62,7 @@
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
+    <PackageReference Include="moq" Version="4.20.69" />
     <PackageReference Include="MSTest.TestAdapter" Version="2.2.10" />
     <PackageReference Include="MSTest.TestFramework" Version="2.2.10" />
     <PackageReference Include="coverlet.collector" Version="3.2.0" />
@@ -69,6 +70,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\SmugMugCore.Net\SmugMugCore.Net.csproj" />
+    <ProjectReference Include="..\SmugMugCoreSync\SmugMugCoreSync.csproj" />
   </ItemGroup>
 
 </Project>
